Clean merge mess, correct some gcc warns
