+++ include/video.hpp
@@ -1,0 +1,11 @@
+#ifndef VIDEO_HPP
+#define VIDEO_HPP
+
+#include <string>
+#include "quantif.hpp"
+
+using namespace std;
+
+void doVid(string source);
+
+#endif /* end of include guard: VIDEO_HPP */
